chore: fix tests
